+++ src/logics/SimpleLogic.ts
@@ -27,6 +27,7 @@ export class SimpleLogic extends AbstractRoutingLogic<"simple"> {
       description: "Simply pull and push tokens directly from your wallet.",
       mgv: params.mgv,
       address: ethers.constants.AddressZero,
+      approvalType: "ERC20",
     });
   }