+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next version
 
+# 2.0.5-44
+
 - Add Smart kandels to SDK
 - Changed `onAave` params to `type` params
 - Added monoswap and thruster routing logics